+++ lib/rails/generators/sass_scaffold.rb
@@ -8,9 +8,8 @@ module Sass
         dir = ::Rails::Generators::ScaffoldGenerator.source_root
         file = File.join(dir, "scaffold.css")
         converted_contents = ::Sass::CSS.new(File.read(file)).render(syntax)
-        create_file "app/assets/stylesheets/scaffolds.css.#{syntax}", converted_contents
+        create_file "app/assets/stylesheets/scaffolds.#{syntax}", converted_contents
       end
     end
   end
 end
-