+++ lib/sass/rails/template.rb
@@ -3,6 +3,9 @@ require "sprockets/sass_template"
 module Sass
   module Rails
     class SassTemplate < Sprockets::SassTemplate
+      def self.default_mime_type
+        'text/css'
+      end
 
       def evaluate(context, locals, &block)
         cache_store = Sprockets::SassCacheStore.new(context.environment)
@@ -12,8 +15,8 @@ module Sass
           :line => line,
           :syntax => syntax,
           :cache_store => cache_store,
-          :importer => SassImporter.new(context.pathname.to_s),
-          :load_paths => context.environment.paths.map { |path| SassImporter.new(path.to_s) },
+          :importer => importer_class.new(context.pathname.to_s),
+          :load_paths => context.environment.paths.map { |path| importer_class.new(path.to_s) },
           :sprockets => {
             :context => context,
             :environment => context.environment
@@ -32,13 +35,15 @@ module Sass
         context.__LINE__ = e.sass_backtrace.first[:line]
         raise e
       end
+
+      private
+
+      def importer_class
+        SassImporter
+      end
     end
 
     class ScssTemplate < SassTemplate
-      def self.default_mime_type
-        'text/css'
-      end
-
       def syntax
         :scss
       end