+++ test/fixtures/alternate_config_project/config/environments/production.rb
@@ -27,7 +27,7 @@ AlternateConfigProject::Application.configure do
   # config.force_ssl = true
 
   # See everything in the log (default is :info)
-  # config.log_level = :debug
+  config.log_level = :info
 
   # Use a different logger for distributed setups
   # config.logger = SyslogLogger.new