+++ test/fixtures/sass_project/config/environments/test.rb
@@ -7,6 +7,9 @@ ScssProject::Application.configure do
   # and recreated between test runs.  Don't rely on the data there!
   config.cache_classes = true
 
+  # Do not eager load code on boot.
+  config.eager_load = false
+
   # Configure static asset server for tests with Cache-Control for performance
   config.serve_static_assets = true
   config.static_cache_control = "public, max-age=3600"