+++ test/sass_rails_test.rb
@@ -8,57 +8,81 @@ class SassRailsTest < Sass::Rails::TestCase
 
   test 'style config item is honored in development mode' do
     within_rails_app 'alternate_config_project' do
-      runcmd "ruby script/rails runner 'puts Rails.application.config.sass.style'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'development'}
+      runner 'development' do
+        "puts Rails.application.config.sass.style"
+      end
+
       assert_output /compact/
     end
   end
 
   test 'style config item is not honored if environment is not development' do
     within_rails_app 'alternate_config_project' do
-      runcmd "ruby script/rails runner 'p Rails.application.config.sass.style'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'production'}
-      assert_equal 'nil', $last_ouput.chomp
+      runner 'production' do
+        "p Rails.application.config.sass.style"
+      end
+
+      assert_equal 'nil', $last_output.chomp
     end
   end
 
   test 'css_compressor config item is not honored in development mode' do
     within_rails_app 'alternate_config_project' do
-      runcmd "ruby script/rails runner 'p Rails.application.config.assets.css_compressor'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'development'}
-      assert_equal 'nil', $last_ouput.chomp
+      runner 'development' do
+        "p Rails.application.config.assets.css_compressor"
+      end
+
+      assert_equal 'nil', $last_output.chomp
     end
   end
 
   test 'css_compressor config item is honored if environment is not development' do
     within_rails_app 'alternate_config_project' do
-      runcmd "ruby script/rails runner 'puts Rails.application.config.assets.css_compressor'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'production'}
+      runner 'production' do
+        "puts Rails.application.config.assets.css_compressor"
+      end
+
       assert_output /yui/
     end
   end
 
   test 'sass uses expanded style by default in development mode' do
     within_rails_app 'scss_project' do
-      runcmd "ruby script/rails runner 'puts Rails.application.config.sass.style'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'development'}
+      runner 'development' do
+        "puts Rails.application.config.sass.style"
+      end
+
       assert_output /expanded/
     end
   end
 
   test 'sass not defines compressor in development mode' do
     within_rails_app 'scss_project' do
-      runcmd "ruby script/rails runner 'p Rails.application.config.assets.css_compressor'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'development'}
-      assert_equal 'nil', $last_ouput.chomp
+      runner 'development' do
+        "p Rails.application.config.assets.css_compressor"
+      end
+
+      assert_equal 'nil', $last_output.chomp
     end
   end
 
   test 'sass defines compressor by default in test mode' do
     within_rails_app 'scss_project' do
-      runcmd "ruby script/rails runner 'puts Rails.application.config.assets.css_compressor'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'test'}
-      assert_equal 'sass', $last_ouput.chomp
+      runner 'test' do
+        "puts Rails.application.config.assets.css_compressor"
+      end
+
+      assert_equal 'sass', $last_output.chomp
     end
   end
 
   test 'sass defines compressor by default in production mode' do
     within_rails_app 'scss_project' do
-      runcmd "ruby script/rails runner 'puts Rails.application.config.assets.css_compressor'", Dir.pwd, true, 'Gemfile', {'RAILS_ENV' => 'production'}
-      assert_equal 'sass', $last_ouput.chomp
+      runner 'production' do
+        "puts Rails.application.config.assets.css_compressor"
+      end
+
+      assert_equal 'sass', $last_output.chomp
     end
   end
 
@@ -68,7 +92,7 @@ class SassRailsTest < Sass::Rails::TestCase
   end
 
   test 'sass imports work correctly' do
-    css_output = sprockets_render('scss_project', 'application.css.scss')
+    css_output = sprockets_render('scss_project', 'application.scss')
     assert_match /main/,                     css_output
     assert_match /top-level/,                css_output
     assert_match /partial-sass/,             css_output
@@ -91,7 +115,7 @@ class SassRailsTest < Sass::Rails::TestCase
 
   test 'globbed imports work when new file is added' do
     project = 'scss_project'
-    filename = 'application.css.scss'
+    filename = 'application.scss'
 
     within_rails_app(project) do |tmpdir|
       asset_output(filename)
@@ -107,7 +131,7 @@ class SassRailsTest < Sass::Rails::TestCase
   end
 
   test 'sass asset paths work' do
-    css_output = sprockets_render('scss_project', 'application.css.scss')
+    css_output = sprockets_render('scss_project', 'application.scss')
     assert_match %r{asset-path:\s*"/assets/rails.png"},                           css_output, 'asset-path:\s*"/assets/rails.png"'
     assert_match %r{asset-url:\s*url\(/assets/rails.png\)},                       css_output, 'asset-url:\s*url\(/assets/rails.png\)'
     assert_match %r{image-path:\s*"/assets/rails.png"},                           css_output, 'image-path:\s*"/assets/rails.png"'