+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG
 
+## unreleased
+
+* Register scss and sass extensions for rake notes
+
+## 4.0.5 (Nov 25, 2014)
+
+* Make possible to use sprockets 2.12.
+
+## 4.0.4 (Oct 29, 2014)
+
+* Make possible to use any sprockets version in the 2.11 series.
+* Require at least sass 3.2.2.
+
+## 4.0.3 (Apr 4, 2014)
+
+* Make possible to use sprockets-rails 2.1.
+
+## 4.0.2 (Mar 13, 2014)
+
+* Lock sprockets version to <= 2.11.0. Fixes #191.
+
 ## 4.0.1 (Oct 15, 2013)
 
 * Remove Post Processors from asset evaluation.